feat: add transclusion snippet and fix zetel/dezetelify script names
This commit introduces a new snippet `trans` to UltiSnips for easy transclusion of files using the `.include` directive.
It also renames the script names to follow the convention in by the `pream` snippet and `IndepthWikiEnterNormal` function, changing `myzetelify` and `mydezetelify` to `my_zetelify` and `my_dezetelify` respectively.
Finally, it adds an autocommand to remove the block cursor on VimEnter,
due to a zshrc problem.